✨ feature: finish the CRUD system with some messy code

+++ front-end-project/peopleFetchPromise.js
@@ -1,49 +1,73 @@
 const addPersonPromise = () => {
-    var newPerson = document.getElementById("iPerson").value;
-    const slug = newPerson.slice(0,3).toLowerCase();
-
+    let newPerson = addPersonInputElement.value;
+    if (newPerson === "")
+    {
+        alert("Person field is empty!")
+        return;
+    }
+    
     fetch(PEOPLE_ENDPOINT, {
         method: "POST",
         headers: {"Content-Type": "application/json"},
         body: JSON.stringify({
-            id: slug,
+            id: create_UUID(),
             name: newPerson,
+            slug: newPerson.slice(0,3).toLowerCase(),
         })
     })
-    .then(resp => getAllPeoplePromise())
-    document.getElementById("iPerson").value = "";
-}
+    .then(resp => updateAllPeopleRenders())
 
-const deletePersonPromise = (id) => {
-    fetch(PEOPLE_ENDPOINT + `/${id}`, {
-        method: "DELETE",        
-    }).then(resp => getAllPeoplePromise())
+    clearAddPersonInputElement();
 }
 
 const getAllPeoplePromise = () => {
-    const promise = fetch(PEOPLE_ENDPOINT, {
+    return fetch(PEOPLE_ENDPOINT, {
         method: "GET"
     }).then(
         resp => resp.json()
-    ).then(data => updateAllPeopleData(data))
+    )
+}
+
+const updateAllPeopleRenders = () => {
+    getAllPeoplePromise()
+        .then(data => {
+            renderAllPeopleList(data);
+            renderDropdownOptionsPeopleList(data);
+        })
+}
+
+const deletePersonPromise = (id) => {
+    getAllTasksPromise().then(data => { 
+        let personTasks = data.filter(x => x.people_id === id);
+        if (personTasks.length === 0)
+            fetch(PEOPLE_ENDPOINT + `/${id}`, {
+                method: "DELETE",        
+            }).then(resp => updateAllPeopleRenders())
+        else {
+            if (confirm(`There are some tasks for this person. Want to delete all?`))
+            {
+                deletePersonTasks(id);
+                fetch(PEOPLE_ENDPOINT + `/${id}`, {
+                    method: "DELETE",        
+                }).then(resp => updateAllPeopleRenders());
+            }
+        }
+    });
+
+    
+}
+
+const insertPeopleWithEnterKey = () => {
+    addPersonInputElement.addEventListener("keypress", (event) => {
+        if (event.key === "Enter")
+            addPersonPromise()
+    })
 }
 
 
-const updateAllPeopleData = (data) => {
-    var container = document.getElementById("list-people");
-    let html = data.map((person) => `<p style='margin:1px; position: relative'>${person.name} 
-    <span style='text-decoration: underline; position: absolute; right: 0px; cursor: pointer' onclick="deletePersonPromise('${person.id}')">delete</span>
-    </p>`);
-    container.innerHTML = html.join("");
-}
+updateAllPeopleRenders();
+insertPeopleWithEnterKey();
 
 
-
-getAllPeoplePromise();
-
-document.getElementById("iPerson").addEventListener("keypress", (event) => {
-    if (event.key === "Enter")
-        addPersonPromise()
-})
